requirements-dev.txt
# Development and Testing Dependencies
# Install with: pip install -r requirements-dev.txt

# Include base requirements
-r requirements.txt

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-watch==4.2.0

# Code Coverage
coverage==7.3.4

# Code Quality
flake8==6.1.0
black==23.12.1
isort==5.13.2

# Type Checking
mypy==1.7.1
django-stubs==4.2.7
